pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-cookiejwt"
version = "0.0.1"
description = "JWT authentication using HTTP-only cookies for Django REST Framework"
readme = "README.md"
authors = [
    {name = "Andrey Kalashnikov", email = "andrey.kalashnikov@muehlemann-popp.ch"},
]
license = {text = "MIT"}
keywords = ["django", "jwt", "authentication", "cookies", "rest-framework"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1", 
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.12"
dependencies = [
    "Django>=4.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.2.0",
    "drf-spectacular>=0.28.0",
]

[project.urls]
Homepage = "https://github.com/muehlemann-popp/django-cookiejwt"
Repository = "https://github.com/muehlemann-popp/django-cookiejwt"
Documentation = "https://github.com/muehlemann-popp/django-cookiejwt/wiki"
"Bug Tracker" = "https://github.com/muehlemann-popp/django-cookiejwt/issues"
Changelog = "https://github.com/muehlemann-popp/django-cookiejwt/blob/master/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/django_cookiejwt"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
testpaths = ["tests"]
addopts = "--cov=django_cookiejwt --cov-report=term-missing --cov-report=xml --cov-report=html"
pythonpath = ["."]

[tool.ruff]
line-length = 120

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "tests.settings"
strict_settings = false
ignore_missing_settings = true

[dependency-groups]
dev = [
    "coverage>=7.8.2",
    "django-stubs>=5.2.0",
    "djangorestframework-stubs>=3.16.0",
    "freezegun>=1.5.2",
    "mypy>=1.16.0",
    "pytest>=8.4.0",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.11.1",
    "ruff>=0.11.12",
]
